+++ docs/changelog.md
@@ -30,6 +30,10 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
 
 ## Unreleased
 
+- New rules:
+  - Container Apps:
+    - Check that liveness and readiness health probes use HTTP checks for HTTP-based ingress.
+      [#3714](https://github.com/Azure/PSRule.Rules.Azure/issues/3714)
 - Updated rules:
   - Azure Kubernetes Service:
     - Updated `Azure.AKS.Version` to use `1.33.7` as the minimum version by @BernieWhite.